A table of installed allometric models

allometric_modelsis a model_tbl containing every installed allometric model in allometric. It is available globally on package load. If not, run the install_models function which will install the models and expose the allometric_models table in your session. This table behaves very much like a tibble::tbl_df or a data.frame``, and most functions that work on these will work on allometric_models` as well.

Details

Printing the head of allometric_models, we can see the structure of the data

options(width=10000)
head(allometric_models)
#> # A tibble: 6 x 13
#>   id       component measure  country   region     family     genus  species model      pub_id        family_name covt_name pub_year
#>   <chr>    <chr>     <chr>    <list>    <list>     <chr>      <chr>  <chr>   <list>     <chr>         <list>      <list>       <dbl>
#> 1 3aff0a28 stem      volume   <chr [1]> <chr [1]>  Aceraceae  Acer   <NA>    <FxdEffcM> brackett_1977 <chr [1]>   <chr [2]>     1977
#> 2 26558408 stem      diameter <chr [2]> <chr [3]>  Betulaceae Alnus  rubra   <FxdEffcM> bluhm_2007    <chr [3]>   <chr [4]>     2007
#> 3 6de9245e stem      volume   <chr [1]> <chr [1]>  Betulaceae Alnus  rubra   <FxdEffcM> brackett_1977 <chr [1]>   <chr [2]>     1977
#> 4 d366d64d stem      diameter <chr [2]> <chr [3]>  Betulaceae Alnus  rubra   <FxdEffcM> hibbs_2007    <chr [3]>   <chr [3]>     2007
#> 5 6dad8922 stem      volume   <chr [2]> <chr [10]> Betulaceae Alnus  rubra   <FxdEffcM> poudel_2019   <chr [4]>   <chr [2]>     2019
#> 6 0b7fcf26 stem      volume   <chr [1]> <chr [1]>  Betulaceae Betula <NA>    <FxdEffcM> brackett_1977 <chr [1]>   <chr [2]>     1977

The columns are:

id - A unique ID for the model.

component - The component of the tree that is modeled.

measure - The measure of the component that is modeled.

country - The country or countries from which the model data is from.

region - The region (e.g., state, province, etc.) from which the model data is from.

family, genus, species - The taxonomic specification of the trees that are modeled.

model - The model object itself.

pub_id - A unique ID representing the publication.

family_name - The names of the contributing authors.

covt_name - The names of the covariates used in the model.

pub_year - The publication year.

Basic Searching for Models

Filtering out nested data from a table is slightly more involved than strictly tabular data. Fortunately the unnest_models function allows the user to unnest any set of columns. For example, let's say we wanted to find a model from the author "Hann". To do this, we will unnest the family_name

Advanced Searching for Models

This table is composed primarily of nested data, referred to as list columns. Nested data, while somewhat cumbersome for users that are not experienced with it, is necessary for accommodating the flexible nature of published allometric models. An obvious motivation is that a given model contains multiple authors, covariates, even multiple regions or countries. Nested data allows for the storage of these highly variable structures for each model.