Brackett (1977)

Stem Volume Models

vsa = 10^a * dsob^b * hst^c

vsa [ft^3] - volume of the entire stem, including top and stump

dsob [in] - diameter of the stem, outside bark at breast height

hst [ft] - total height of the stem

Model Parameters

family       genus       species        region     age_cl…¹     a     b     c
Pinaceae     Pseudotsuga  menziesii      coastal    < 140 … -2.66  1.74 1.13 
Pinaceae     Pseudotsuga  menziesii      coastal    >= 80 … -2.71  1.66 1.20 
Pinaceae     Pseudotsuga  menziesii      interior           -2.73  1.74 1.17 
Pinaceae     Tsuga        heterophylla   coastal    < 140 … -2.70  1.84 1.12 
Pinaceae     Tsuga        heterophylla   coastal    >= 80 … -2.66  1.79 1.12 
Pinaceae     Tsuga        heterophylla   interior           -2.57  1.97 0.977
Cupressaceae Thuja        plicata        coastal    < 140 … -2.44  1.72 1.05 
Cupressaceae Thuja        plicata        coastal    >= 80 … -2.38  1.68 1.04 
Cupressaceae Thuja        plicata        interior           -2.46  1.70 1.07 
Pinaceae     Abies        balsamea       coastal            -2.58  1.81 1.09 
Pinaceae     Abies        balsamea       interior           -2.50  1.86 1.00 
Pinaceae     Picea        stichensis                < 140 … -2.55  1.84 1.04 
Pinaceae     Picea        stichensis                >= 140… -2.70  1.75 1.16 
Pinaceae     Picea        stichensis     interior           -2.54  1.84 1.03 
Pinaceae     Pinus        ponderosa                         -2.73  1.91 1.09 
Pinaceae     Pinus        contorta                          -2.62  1.85 1.09 
Pinaceae     Pinus        monticola                         -2.48  1.87 0.994
Pinaceae     Larix        occidentalis                      -2.62  1.85 1.04 
Cupressaceae Cupressus    nootkatensis                      -2.45  1.74 1.06 
Betulaceae   Alnus        rubra                             -2.67  1.92 1.07 
Aceraceae    Acer                                           -2.77  1.89 1.12 
Salicaceae   Populus      tremuloides                       -2.64  1.95 1.02 
Betulaceae   Betula                                         -2.76  1.91 1.11 
Salicaceae   Populus                                        -2.95  2    1.24 
 with abbreviated variable name ¹​age_class