Hann (2011)

Allometric models from Hann (2011)

Stem Diameter Models

dsib = b_1 * dsob^b_2 * exp(b_3 * (1 - rc)^(0.5))

dsib [in] - diameter of the stem, inside bark at breast height

dsob [in] - diameter of the stem, outside bark at breast height

rc [ft/ft] - crown ratio

Model Parameters

amily       genus       species       b_1   b_2     b_3
inaceae     Pseudotsuga menziesii   0.924 0.989 -0.0341
upressaceae Calocedrus  <NA>        0.879 1     -0.0770
inaceae     Pinus       ponderosa   0.809 1.02   0     
inaceae     Pinus       lambertiana 0.859 1      0     
inaceae     Abies       concolor    0.922 1     -0.0342
inaceae     Abies       grandis     0.922 1     -0.0342

Stump Diameter Models

dui1 = b_1 + b_2 * dsob^b_3 * exp(b_4 * rc^b_5)

dui1 [in] - diameter of the stump, inside bark at one unit

dsob [in] - diameter of the stem, outside bark at breast height

rc [ft/ft] - crown ratio

Model Parameters

amily       genus       species       b_1   b_2 b_3            b_4   b_5
inaceae     Pseudotsuga menziesii   0.150 0.901 1.021366311  0.134  3.68
upressaceae Calocedrus  <NA>        0.452 0.832 1.056002686  0.216  7.00
inaceae     Pinus       ponderosa   0     1.00  0            0      0   
inaceae     Pinus       lambertiana 0     1.04  0            0      0   
inaceae     Abies       concolor    0.393 0.730 1 0978510098 0.121  1   
inaceae     Abies       grandis     0.393 0.730 1 0978510098 0.121  1   

Stem Volume Models

vso = b_1 * (hso/dsob)^(b_2 * (1 - exp(b_3 * dsob^b_4)))^k * exp(b_5 * (hsv/hso)) * dsob^b_7 * dsob^2 * hso

vso [ft^3] - volume of the stem above breast height

hso [ft] - height of the stem above breast height

dsob [in] - diameter of the stem, outside bark at breast height

hsv [ft] - height of the stem to crown base above breast height

Model Parameters

amily       genus       species          b_1   b_2    b_3   b_4    b_5   b_6     b_7     k
inaceae     Pseudotsuga menziesii   0.00124  0.346 -0.566 0.632 -0.152  4.56 -0.0512     1
upressaceae Calocedrus  <NA>        0.000908 0.343 -0.639 1.59   0      0     0          1
inaceae     Pinus       ponderosa   0.00128  0.162  1     0      0      0     0          0
inaceae     Pinus       lambertiana 0.000856 0.388  1     0      0      0     0          0
inaceae     Abies       concolor    0.00133  0.358 -0.755 0.5   -0.262  1     0          1
inaceae     Abies       grandis     0.00133  0.358 -0.755 0.5   -0.262  1     0          1