Huynh et al. (2022)

Allometric models from Huynh et al. (2022)

Tree Biomass Models

bt = alpha * dsob^beta

bt [kg] - tree biomass

dsob [cm] - diameter of the stem, outside bark at breast height

Model Parameters

  equation_no  alpha  beta
1 3           0.0822  2.64

bt = alpha * hst^beta

bt [kg] - tree biomass

hst [m] - total height of the stem

Model Parameters

  equation_no   alpha  beta
1 4           0.00622  3.50

bt = alpha * dsob^beta * hst^beta_1

bt [kg] - tree biomass

dsob [cm] - diameter of the stem, outside bark at breast height

hst [m] - total height of the stem

Model Parameters

  equation_no  alpha  beta beta_1 
1 5           0.0525  2.40 0.38285

bt = alpha * (dsob^2 * hst)^beta

bt [kg] - tree biomass

dsob [cm] - diameter of the stem, outside bark at breast height

hst [m] - total height of the stem

Model Parameters

  equation_no  alpha  beta
1 6           0.0253  1.01

bt = alpha * dsob^beta * rwd^beta_1

bt [kg] - tree biomass

dsob [cm] - diameter of the stem, outside bark at breast height

rwd [kg m-3] - wood specific gravity

Model Parameters

  equation_no  alpha  beta beta_1 
1 7           0.0525  2.40 0.38253

bt = alpha * dsob^beta + hst^beta_1 + rwd^beta_2

bt [kg] - tree biomass

dsob [cm] - diameter of the stem, outside bark at breast height

hst [m] - total height of the stem

rwd [kg m-3] - wood specific gravity

Model Parameters

  equation_no   alpha  beta beta_1  beta_2
1 8           0.00233  2.43 0.30576  0.499

bt = alpha * (dsob^2 * hst * rwd)^beta

bt [kg] - tree biomass

dsob [cm] - diameter of the stem, outside bark at breast height

hst [m] - total height of the stem

rwd [kg m-3] - wood specific gravity

Model Parameters

  equation_no   alpha  beta
1 9           0.00004 0.990

bt = alpha * dsob^beta

bt [kg] - tree biomass

dsob [cm] - diameter of the stem, outside bark at breast height

Model Parameters

  equation_no alpha  beta
1 10          0.106  2.57

bt = alpha * hst^beta

bt [kg] - tree biomass

hst [m] - total height of the stem

Model Parameters

  equation_no   alpha  beta
1 11          0.00027  4.45

bt = alpha * dc^beta

bt [kg] - tree biomass

dc [m] - variable not defined

Model Parameters

  equation_no alpha  beta
1 12           33.2  1.62

bt = alpha * dsob^beta * dc^beta_1

bt [kg] - tree biomass

dsob [cm] - diameter of the stem, outside bark at breast height

dc [m] - variable not defined

Model Parameters

  equation_no alpha  beta beta_1
1 13           2.30  1.07 <NA>  

bt = alpha * dsob^beta * hst^beta_1 * dc^beta_2

bt [kg] - tree biomass

dsob [cm] - diameter of the stem, outside bark at breast height

hst [m] - total height of the stem

dc [m] - variable not defined

Model Parameters

  equation_no  alpha  beta beta_1  beta_2
1 14          0.0515  2.19 0.54648  0.117

bt = alpha * (dsob^2 * hst * dc)^beta

bt [kg] - tree biomass

dsob [cm] - diameter of the stem, outside bark at breast height

hst [m] - total height of the stem

dc [m] - variable not defined

Model Parameters

  equation_no alpha  beta
1 15          0.196 0.680

bt = alpha * dsob^beta * hst^beta_1 * rwd^beta_2 * dc^beta_3

bt [kg] - tree biomass

dsob [cm] - diameter of the stem, outside bark at breast height

hst [m] - total height of the stem

rwd [kg m-3] - wood specific gravity

dc [m] - variable not defined

Model Parameters

  equation_no   alpha  beta beta_1  beta_2 beta_3
1 16          0.00079  2.07 0.69202  0.603  0.180

bt = alpha * (dsob^2 * hst * rwd * dc)^beta

bt [kg] - tree biomass

dsob [cm] - diameter of the stem, outside bark at breast height

hst [m] - total height of the stem

rwd [kg m-3] - wood specific gravity

dc [m] - variable not defined

Model Parameters

  equation_no   alpha  beta
1 17          0.00156 0.699

bt = alpha * vc^beta

bt [kg] - tree biomass

vc [cm3] - crown volume

Model Parameters

  equation_no alpha  beta
1 18           15.4 0.539

bt = alpha * dsob^beta * vc^beta_1

bt [kg] - tree biomass

dsob [cm] - diameter of the stem, outside bark at breast height

vc [cm3] - crown volume

Model Parameters

  equation_no  alpha  beta beta_1  
1 19          0.0988  2.61 ?0.01109

bt = alpha * dsob^beta * hst^beta_1 * vc^beta_2

bt [kg] - tree biomass

dsob [cm] - diameter of the stem, outside bark at breast height

hst [m] - total height of the stem

vc [cm3] - crown volume

Model Parameters

  equation_no  alpha  beta beta_1 beta_2
1 20          0.0487  2.19 0.5465 0.0391

bt = alpha * (dsob^2 * hst * vc)^beta

bt [kg] - tree biomass

dsob [cm] - diameter of the stem, outside bark at breast height

hst [m] - total height of the stem

vc [cm3] - crown volume

Model Parameters

  equation_no alpha  beta
1 21          0.954 0.383

bt = alpha * dsob^beta * hst^beta_1 * rwd^beta_2 * vc^beta_3

bt [kg] - tree biomass

dsob [cm] - diameter of the stem, outside bark at breast height

hst [m] - total height of the stem

rwd [kg m-3] - wood specific gravity

vc [cm3] - crown volume

Model Parameters

  equation_no   alpha  beta beta_1  beta_2 beta_3
1 22          0.00072  2.07 0.69205  0.603 0.0600

bt = alpha * (dsob^2 * hst * rwd * vc)^beta

bt [kg] - tree biomass

dsob [cm] - diameter of the stem, outside bark at breast height

hst [m] - total height of the stem

rwd [kg m-3] - wood specific gravity

vc [cm3] - crown volume

Model Parameters

  equation_no  alpha  beta
1 23          0.0658 0.389