Krisnawati (2016)

Allometric models from Krisnawati (2016)

Stem Volume Models

vsk = a * dsob^b * hst^c

vsk [m3] - volume of the stem including bark

dsob [cm] - diameter of the stem, outside bark at breast height

hst [m] - total height of the stem

Model Parameters

          a     b     c
1 0.0000636  1.74  1.07

Stem Volume Models

vsa = a * dsob^b * hst^c

vsa [m3] - volume of the entire stem, including top and stump

dsob [cm] - diameter of the stem, outside bark at breast height

hst [m] - total height of the stem

Model Parameters

          a     b     c
1 0.0000542  1.08  1.04

Stem Diameter Models

dsoh = a * dsob^b * hst^c * hsl^d

dsoh [cm] - diameter of the stem, outside bark at a specified height

dsob [cm] - diameter of the stem, outside bark at breast height

hst [m] - total height of the stem

hsl [m] - distance from tip of stem to merchantibility limit

Model Parameters

      a     b      c     d
1  1.38 0.868 -0.639 0.676

Stem Diameter Models

dsih = a * dsob^b * hst^c * hsl^d

dsih [cm] - diameter of the stem, inside bark at a specified height

dsob [cm] - diameter of the stem, outside bark at breast height

hst [m] - total height of the stem

hsl [m] - distance from tip of stem to merchantibility limit

Model Parameters

  volume_type                                              a     b      c     d
1 merchantable volume outside bark to specified height  1.26 0.889 -0.621 0.643

Stem Volume Models

vsm = a * dsob^b * hst^c * (hst^d - (e * dsoh^f * dsob^g * hst^h))

vsm [m3] - volume of the merchantable portion of the stem

dsob [cm] - diameter of the stem, outside bark at breast height

hst [m] - total height of the stem

dsoh [cm] - diameter of the stem, outside bark at a specified height

Model Parameters

          a     b     c     d     e     f     g     h
1 0.0000636  1.74 -1.28  2.35 0.326  3.48 -3.02  2.22