A table of installed allometric models

allometric_modelsis a model_tbl containing every installed allometric model in allometric. It is available globally on package load. If not, run the install_models function which will install the models and expose the allometric_models table in your session. This table behaves very much like a tibble::tbl_df or a data.frame``, and most functions that work on these will work on allometric_models` as well.

Details

Printing the head of allometric_models, we can see the structure of the data

head(allometric_models)
#> # A tibble: 6 x 13
#>   id       component measure country   region    family        genus       species     model      pub_id        family_name covt_name pub_year
#>   <chr>    <chr>     <chr>   <list>    <list>    <chr>         <chr>       <chr>       <list>     <chr>         <list>      <list>       <dbl>
#> 1 18c3398d stem      volume  <chr [1]> <chr [1]> Aceraceae     Acer        <NA>        <FxdEffcM> brackett_1977 <chr [1]>   <chr [2]>     1977
#> 2 94c9f45b stem      volume  <chr [1]> <chr [4]> Altingiaceae  Liquidambar styraciflua <FxdEffcM> hahn_1991     <chr [2]>   <chr [2]>     1991
#> 3 90269872 stem      volume  <chr [1]> <chr [4]> Altingiaceae  Liquidambar styraciflua <FxdEffcM> hahn_1991     <chr [2]>   <chr [2]>     1991
#> 4 f814022d stem      volume  <chr [1]> <chr [1]> Anacardiaceae Tapirira    guianensis  <FxdEffcM> vibrans_2015  <chr [4]>   <chr [2]>     2015
#> 5 6938ab84 tree      biomass <chr [1]> <chr [1]> Aquifoliaceae Ilex        canariensis <FxdEffcM> montero_2005  <chr [3]>   <chr [1]>     2005
#> 6 b994e61a stem      biomass <chr [1]> <chr [1]> Aquifoliaceae Ilex        canariensis <FxdEffcM> montero_2005  <chr [3]>   <chr [1]>     2005

The columns are:

id - A unique ID for the model.

component - The component of the tree that is modeled.

measure - The measure of the component that is modeled.

country - The country or countries from which the model data is from.

region - The region or regions (e.g., state, province, etc.) from which the model data is from.

family, genus, species - The taxonomic specification of the trees that are modeled.

model - The model object itself.

pub_id - A unique ID representing the publication.

family_name - The names of the contributing authors.

covt_name - The names of the covariates used in the model.

pub_year - The publication year.

Basic Searching for Models

Filtering out nested data from a table is slightly more involved than strictly tabular data. Fortunately the unnest_models function allows the user to unnest any set of columns. For example, let's say we wanted to find a model from the author "Hann". To do this, we will unnest the family_name column using unnest_models, then filter the resulting dataframe using dplyr::filter.

unnest_family <- allometric_models %>% unnest_models('family_name')

unnest_family %>% dplyr::filter(family_name == "Hann")
#> # A tibble: 40 x 13
#>    id       component measure  country   region    family       genus            species      model      pub_id       family_name covt_name pub_year
#>    <chr>    <chr>     <chr>    <list>    <list>    <chr>        <chr>            <chr>        <list>     <chr>        <chr>       <list>       <dbl>
#>  1 651aaa4a crown     ratio    <chr [1]> <chr [1]> Cupressaceae Calocedrus       decurrens    <FxdEffcM> ritchie_1987 Hann        <chr [4]>     1987
#>  2 22003acc stem      diameter <chr [1]> <chr [1]> Cupressaceae Calocedrus       <NA>         <FxdEffcM> hann_2011    Hann        <chr [2]>     2011
#>  3 e5d4894e stump     diameter <chr [1]> <chr [1]> Cupressaceae Calocedrus       <NA>         <FxdEffcM> hann_2011    Hann        <chr [2]>     2011
#>  4 91e2d29d stem      volume   <chr [1]> <chr [1]> Cupressaceae Calocedrus       <NA>         <FxdEffcM> hann_2011    Hann        <chr [3]>     2011
#>  5 53162fb2 crown     ratio    <chr [1]> <chr [1]> Ericaceae    Arbutus          menziesii    <FxdEffcM> ritchie_1987 Hann        <chr [3]>     1987
#>  6 c0a86b32 crown     ratio    <chr [1]> <chr [1]> Fagaceae     Chrysolepis      chrysophylla <FxdEffcM> ritchie_1987 Hann        <chr [4]>     1987
#>  7 ca55f6c1 crown     ratio    <chr [1]> <chr [1]> Fagaceae     Notholithocarpus densiflorus  <FxdEffcM> ritchie_1987 Hann        <chr [1]>     1987
#>  8 3da0e21e crown     ratio    <chr [1]> <chr [1]> Fagaceae     Quercus          chrysolepis  <FxdEffcM> ritchie_1987 Hann        <chr [1]>     1987
#>  9 77e11da9 crown     ratio    <chr [1]> <chr [1]> Fagaceae     Quercus          garryana     <FxdEffcM> ritchie_1987 Hann        <chr [1]>     1987
#> 10 cca2d9dc crown     ratio    <chr [1]> <chr [1]> Fagaceae     Quercus          kelloggii    <FxdEffcM> ritchie_1987 Hann        <chr [3]>     1987
#> # ... with 30 more rows

Any column or any combination of columns can be unnested, which allows for basic filtering of models using dplyr::filter.

Advanced Searching for Models

Nested data can be searched directly without the use of unnest_models. This requires the use of purrr::map_lgl which is used to determine truthiness of expressions for each element in a list column. Before beginning, it is helpful to know that purrr::map_lgl returns a list of TRUE/FALSE values as it "maps" over a list of input.

Finding Contributing Authors

Using this function, we can recreate the previous example, finding models that had 'Hann' as a contributing author.

hann_models <- dplyr::filter(
 allometric_models,
 purrr::map_lgl(family_name, ~ 'Hann' %in% .)
)

head(hann_models)
#> # A tibble: 6 x 13
#>   id       component measure  country   region    family       genus       species      model      pub_id       family_name covt_name pub_year
#>   <chr>    <chr>     <chr>    <list>    <list>    <chr>        <chr>       <chr>        <list>     <chr>        <list>      <list>       <dbl>
#> 1 651aaa4a crown     ratio    <chr [1]> <chr [1]> Cupressaceae Calocedrus  decurrens    <FxdEffcM> ritchie_1987 <chr [2]>   <chr [4]>     1987
#> 2 22003acc stem      diameter <chr [1]> <chr [1]> Cupressaceae Calocedrus  <NA>         <FxdEffcM> hann_2011    <chr [1]>   <chr [2]>     2011
#> 3 e5d4894e stump     diameter <chr [1]> <chr [1]> Cupressaceae Calocedrus  <NA>         <FxdEffcM> hann_2011    <chr [1]>   <chr [2]>     2011
#> 4 91e2d29d stem      volume   <chr [1]> <chr [1]> Cupressaceae Calocedrus  <NA>         <FxdEffcM> hann_2011    <chr [1]>   <chr [3]>     2011
#> 5 53162fb2 crown     ratio    <chr [1]> <chr [1]> Ericaceae    Arbutus     menziesii    <FxdEffcM> ritchie_1987 <chr [2]>   <chr [3]>     1987
#> 6 c0a86b32 crown     ratio    <chr [1]> <chr [1]> Fagaceae     Chrysolepis chrysophylla <FxdEffcM> ritchie_1987 <chr [2]>   <chr [4]>     1987
nrow(hann_models)
#> [1] 40

Picking apart the above code block, we see that we are using the standard dplyr::filter function on the allometric_models dataframe. The second argument is a call using purrr:map_lgl, which will map over each list (contained as elements in the family_names column). The second argument to this function, ~ 'Hann' %in% . is itself a function that checks if 'Hann' is in the current list. Imagine we are marching down each row of allometric_models, . represents the element of family_names we are considering, which is itself a list of author names.

Finding First Authors

Maybe we are only interested in models where 'Hann' is the first author. Using a simple modification we can easily do this.

hann_first_author_models <- dplyr::filter(
  allometric_models,
  purrr::map_lgl(family_name, ~ 'Hann' == .[[1]])
)

head(hann_first_author_models)
#> # A tibble: 6 x 13
#>   id       component measure  country   region    family       genus      species  model      pub_id    family_name covt_name pub_year
#>   <chr>    <chr>     <chr>    <list>    <list>    <chr>        <chr>      <chr>    <list>     <chr>     <list>      <list>       <dbl>
#> 1 22003acc stem      diameter <chr [1]> <chr [1]> Cupressaceae Calocedrus <NA>     <FxdEffcM> hann_2011 <chr [1]>   <chr [2]>     2011
#> 2 e5d4894e stump     diameter <chr [1]> <chr [1]> Cupressaceae Calocedrus <NA>     <FxdEffcM> hann_2011 <chr [1]>   <chr [2]>     2011
#> 3 91e2d29d stem      volume   <chr [1]> <chr [1]> Cupressaceae Calocedrus <NA>     <FxdEffcM> hann_2011 <chr [1]>   <chr [3]>     2011
#> 4 88fa6033 stem      diameter <chr [1]> <chr [1]> Pinaceae     Abies      concolor <FxdEffcM> hann_2011 <chr [1]>   <chr [2]>     2011
#> 5 4c61fd99 stump     diameter <chr [1]> <chr [1]> Pinaceae     Abies      concolor <FxdEffcM> hann_2011 <chr [1]>   <chr [2]>     2011
#> 6 79fd8e44 stem      volume   <chr [1]> <chr [1]> Pinaceae     Abies      concolor <FxdEffcM> hann_2011 <chr [1]>   <chr [3]>     2011
nrow(hann_first_author_models)
#> [1] 20

We can see that 'Hann' is the first author for 20 models in this package.

Finding a Model with Specific Data Requirements

We can even check for models that contain certain types of data requirements. For example, the following block finds diameter-height models, specifically models that use diameter outside bark at breast height as the only covariate. The utility here is obvious, since many inventories are vastly limited by their available tree measurements.

dia_ht_models <- dplyr::filter(
    allometric_models,
    measure == 'height',
    purrr::map_lgl(covt_name, ~ length(.)==1 & .[[1]] == 'dsob'),
)

nrow(dia_ht_models)
#> [1] 34

Breaking this down, we have the first condition measure=='height' selecting only models concerned with heights as a response variable. The second line maps over each element of the covt_name column, which is a character vector. The . represents a given character vector for that row. First, we ensure that the vector is only one element in size using length(.)==1, then we ensure that the first (and only) element of this vector is equal to 'dsob', (diameter outside bark at breast height). In this case, 34 are available in the package.

Finding a Model for a Region

By now the user should be sensing a pattern. We can apply the exact same logic as the Finding Contributing Authors section to find all models developed using data from US-OR

us_or_models <- dplyr::filter(
    allometric_models,
    purrr::map_lgl(region, ~ "US-OR" %in% .),
)

nrow(us_or_models)
#> [1] 159

We can see that 159 allometric models are defined for the state of Oregon, US.